Fixed Silverlight support
Since Silverlight support was added to AVIM, Silverlight broke KeyboardEventArgs.Handled and essentially stopped firing KeyUp events. This change expands upon the hacky VIQR Silverlight support introduced way back when. For VIQR, AVIM now eats the inserted character and interprets it asynchronously after KeyDown instead of on KeyUp. Additionally, changes on KeyDown are committed asynchronously. This revised implementation should be compatible with a variety of keyboard layouts, although there is some flickering in VIQR as the character is inserted and eaten back up.