Restored <em:unpack> to install manifest
The Mozilla Add-ons validator says reading the extension’s contents out of a .jar is slow when the extension remains packed up in an .xpi. A .jar was required before Firefox 4, so we need to keep the overall extension unpacked until we drop support for Firefox 2.x and 3.x.

This reverts much of commit ea55e57133d1b8917cac79e1e1dc754f87e2eee3.