First pass on restartlessness
Began migrating from declarative XUL overlays to data-driven but nonetheless procedural JavaScript. 😢 Registered resource: aliases to manually import transformer.js as a JavaScript code module instead of an XPCOM component and the preferences file as a subscript, all while hopefully maintaining Firefox 2 compatibility.

The upside is that AVIM will be restartless, and the redundant menu definitions can be combined.

Fixes #9.