Added menu item for inserting syllable break
A new command looks for the currently edited element and causes the frame script to insert a syllable break in its editor at the current selection. The command is hooked up to a menu item that displays the key binding (in all the overlays except the one for Komodo; syllable breaking isn’t supported in SciMoz). Hopefully makes the syllable break feature more discoverable.