Highlight last syllable when inserting syllable break
When inserting a syllable break, highlight the syllable to the left of the insertion point with an overlay. The overlay is inserted into an HTML document as anonymous content and into an XUL document as ordinary content. It fades out and is removed from the document after a second delay. The overlay is styled using a new user agent stylesheet.

Fixes #153.